+++ lib/neo4j/core/schema.rb
@@ -12,18 +12,49 @@ module Neo4j
         result = query('CALL db.indexes()', {}, skip_instrumentation: true)
 
         result.map do |row|
-          label, property = row.description.match(/INDEX ON :([^\(]+)\(([^\)]+)\)/)[1, 2]
-          { type: row.type.to_sym, label: label.to_sym, properties: [property.to_sym], state: row.state.to_sym }
+          { type: row.type.to_sym, label: label(result, row), properties: properties(row), state: row.state.to_sym }
         end
       end
 
       def constraints
         result = query('CALL db.indexes()', {}, skip_instrumentation: true)
 
-        result.select { |row| row.type == 'node_unique_property' }.map do |row|
-          label, property = row.description.match(/INDEX ON :([^\(]+)\(([^\)]+)\)/)[1, 2]
-          { type: :uniqueness, label: label.to_sym, properties: [property.to_sym] }
+        result.select(&method(v4?(result) ? :v4_filter : :v3_filter)).map do |row|
+          { type: :uniqueness, label: label(result, row), properties: properties(row) }
         end
+      end
+
+      def named_constraints
+        result = query('CALL db.constraints()', {}, skip_instrumentation: true)
+        result.columns.include?(:name) ? result : []
+      end
+
+      private
+
+      def v4_filter(row)
+        row.uniqueness == 'UNIQUE'
+      end
+
+      def v3_filter(row)
+        row.type == 'node_unique_property'
+      end
+
+      def label(result, row)
+        (v34?(result) ? row.label : (v4?(result) ? row.labelsOrTypes : row.tokenNames).first).to_sym
+      end
+
+      def v4?(result)
+        return @v4 unless @v4.nil?
+        @v4 = result.columns.include?(:labelsOrTypes)
+      end
+
+      def v34?(result)
+        return @v34 unless @v34.nil?
+        @v34 = result.columns.include?(:label)
+      end
+
+      def properties(row)
+        row.properties.map(&:to_sym)
       end
     end
   end