+++ spec/migration_files/migrations/1234567890_rename_john_jack.rb
@@ -1,6 +1,6 @@
 class RenameJohnJack < Neo4j::Migrations::Base
   def up
-    execute 'MATCH (u:`User`) WHERE u.name = {name} SET u.name = {new_name}',
+    execute 'MATCH (u:`User`) WHERE u.name = $name SET u.name = $new_name',
             name: 'John', new_name: 'Jack'
   end