+++ spec/migration_files/transactional_migrations/1231231231_failing_migration.rb
@@ -1,6 +1,6 @@
 class FailingMigration < Neo4j::Migrations::Base
   def up
-    execute 'MATCH (u:`User`) WHERE u.name = {name} SET u.name = {new_name}',
+    execute 'MATCH (u:`User`) WHERE u.name = $name SET u.name = $new_name',
             name: 'Joe', new_name: 'Jack'
     execute 'CREATE (n:`Contact` {phone: "123123"})'
   end